chore: add remote config for growthbook
